refactor: 抽取 typescript 类型
