+++ types/lesson.d.ts
@@ -1,0 +1,25 @@
+export interface Lesson {
+  id: number
+  titleEn: string
+  titleZh: string
+  sentences: Sentence[]
+  audioUrl: string
+}
+
+export interface Sentence {
+  lessonId: number
+  sentenceId: number
+  startAt: number
+  stopAt: number
+  en: string
+  zh: string
+}
+
+export interface SentenceInfo {
+  isAnswerVisible: boolean
+  isMarked: boolean
+  inputText: string
+  sentence: Sentence
+  diffChanges?: Diff.Change[]
+  audioUrl: string
+}
